switching over to redux toolkit
