+++ package.json
@@ -14,6 +14,7 @@
 		"@fortawesome/free-regular-svg-icons": "5.4.1",
 		"@fortawesome/free-solid-svg-icons": "5.3.1",
 		"@fortawesome/react-fontawesome": "0.1.3",
+		"@reduxjs/toolkit": "1.1.0",
 		"@welldone-software/why-did-you-render": "2.5.0",
 		"angular": "1.5.11",
 		"babel-core": "7.0.0-bridge.0",
