+++ src/store.js
@@ -1,50 +1,34 @@
-import { combineReducers } from "redux";
+import { configureStore, getDefaultMiddleware } from "@reduxjs/toolkit";
 import { createBrowserHistory } from "history";
-import { applyMiddleware, compose, createStore } from "redux";
 import { connectRouter, routerMiddleware } from "connected-react-router";
-import thunk from "redux-thunk";
-import localUser from "./features/users/localUserReducer";
-import localRoomMembers from "./features/users/localRoomMembersReducer";
-import userSettings from "./features/settings/userSettingsReducer";
-import users from "./features/users/usersReducer";
-import rooms from "./features/room/roomReducer";
-import messages from "./features/message/messagesReducer";
-import chatInput from "./features/input/chatInputReducer";
-import emoticonPicker from "./features/emoticon/emoticonPickerReducer";
-import messageControls from "./features/messageControls/messageControlsReducer";
-import imagePick from "./features/imagePick/imagePickReducer";
-import log from "./features/chat/logReducer";
-import imageUpload from "./features/imageUpload/imageUploadReducer.js";
-
-const rootReducer = combineReducers({
-	localUser,
-	localRoomMembers,
-	userSettings,
-	users,
-	rooms,
-	messages,
-	chatInput,
-	emoticonPicker,
-	messageControls,
-	imagePick,
-	log,
-	imageUpload
-});
+// import thunk from "redux-thunk";
+import rootReducer from "./rootReducer";
 
 const history = createBrowserHistory();
 
-const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
+const store = configureStore({
+	reducer: connectRouter(history)(rootReducer),
+	middleware: [routerMiddleware(history), ... getDefaultMiddleware()]
+});
 
-const store = createStore(
-	connectRouter(history)(rootReducer), // new root reducer with router state
-	composeEnhancers(
-		applyMiddleware(
-			routerMiddleware(history), // for dispatching history actions
-			thunk
-			// ... other middlewares ...
-		)
-	)
-);
+//
+// const store = createStore(
+// 	connectRouter(history)(rootReducer), // new root reducer with router state
+// 	composeEnhancers(
+// 		applyMiddleware(
+// 			routerMiddleware(history), // for dispatching history actions
+// 			thunk
+// 			// ... other middlewares ...
+// 		)
+// 	)
+// );
+
+if (process.env.NODE_ENV === "development" && module.hot) {
+	module.hot.accept("./rootReducer", () => {
+		const newRootReducer = require("./rootReducer").default;
+		store.replaceReducer(newRootReducer);
+	});
+}
 
 const dispatch = store.dispatch;
